Materialize DiffusionGemma processor video outputs
